+++ CHANGES.md
@@ -9,6 +9,7 @@
 - Adding unit tests for `annotate-sv` command (#33)
 - Writing out gzip-ed files if output file name ends in `.gz` (#11)
 - Do not write out asterisk alleles (#18)
+- Adding tests for GRCh38 (#22)
 
 ## v0.21